Rename directory MD->Markdown
